firmware: Enable swipe gesture for octave change.
Collect finger position delta values in a buffer and sum them to detect
swiping gestures.

This is much more convenient than the old way of having static mapping
of the ribbon as octave zones. Old position information is not
conflicting and can be used for some other mapping if needed.